+++ src/Component/CommGS/InitAnt.hpp
@@ -1,5 +1,5 @@
 #pragma once
 
-#include <Component/CommGS/ANT.h>
+#include <Component/CommGS/Antenna.hpp>
 
 ANT InitANT(int ant_id, const std::string fname);
